+++ src/Model/Table/AppTable.php
@@ -9,7 +9,6 @@ use Cake\ORM\Marshaller;
 use Cake\ORM\Table;
 use Cake\Datasource\FactoryLocator;
 use Cake\Utility\Hash;
-use Cake\Utility\Security;
 use Cake\Validation\Validation;
 use Cake\Validation\Validator;
 
@@ -178,197 +177,6 @@ class AppTable extends Table
             }
         }
         return true;
-    }
-
-    public function getProductIdentifierField()
-    {
-        return 'SUBSTRING(SHA1(CONCAT(Products.id_product, "' .  Security::getSalt() . '", "product")), 1, 4)';
-    }
-
-    /**
-     * @return string
-     */
-    protected function getFieldsForProductListQuery()
-    {
-        $fields = "Products.id_product,
-                Products.name, Products.description_short, Products.description, Products.unity, Products.price, Products.created, Products.is_stock_product,
-                Products.delivery_rhythm_type, Products.delivery_rhythm_count, Products.delivery_rhythm_first_delivery_day, Products.delivery_rhythm_order_possible_until,
-                Products.delivery_rhythm_send_order_list_weekday, Products.delivery_rhythm_send_order_list_day,
-                Deposits.deposit,
-                Images.id_image,
-                Manufacturers.id_manufacturer, Manufacturers.name as ManufacturersName,
-                Manufacturers.timebased_currency_enabled, Manufacturers.no_delivery_days, Manufacturers.stock_management_enabled,
-                Units.price_per_unit_enabled, Units.price_incl_per_unit, Units.name as unit_name, Units.amount as unit_amount, Units.quantity_in_units,
-                Units.purchase_price_incl_per_unit,
-                Taxes.rate as taxRate,
-                StockAvailables.quantity, StockAvailables.quantity_limit, StockAvailables.always_available";
-
-        if (Configure::read('appDb.FCS_TIMEBASED_CURRENCY_ENABLED')) {
-            $fields .= ", Manufacturers.timebased_currency_max_percentage, Manufacturers.timebased_currency_max_credit_balance";
-        }
-
-        if (Configure::read('appDb.FCS_SELF_SERVICE_MODE_FOR_STOCK_PRODUCTS_ENABLED')) {
-            $fields .= ", " . $this->getProductIdentifierField() . " as ProductIdentifier";
-            $fields .= ", ProductBarcodes.barcode as ProductBarcode";
-            $fields .= ", ProductAttributeBarcodes.barcode as ProductAttributeBarcode, ProductAttributes.id_product_attribute as ProductAttributeId";
-        }
-
-        $fields .= " ";
-        return $fields;
-    }
-
-    /**
-     * @return string
-     */
-    protected function getJoinsForProductListQuery()
-    {
-        $sql = "LEFT JOIN ".$this->tablePrefix."stock_available StockAvailables ON Products.id_product = StockAvailables.id_product
-                LEFT JOIN ".$this->tablePrefix."images Images ON Images.id_product = Products.id_product
-                LEFT JOIN ".$this->tablePrefix."deposits Deposits ON Products.id_product = Deposits.id_product
-                LEFT JOIN ".$this->tablePrefix."units Units ON Products.id_product = Units.id_product
-                LEFT JOIN ".$this->tablePrefix."tax Taxes ON Products.id_tax = Taxes.id_tax
-                LEFT JOIN ".$this->tablePrefix."manufacturer Manufacturers ON Manufacturers.id_manufacturer = Products.id_manufacturer ";
-
-        if (Configure::read('appDb.FCS_SELF_SERVICE_MODE_FOR_STOCK_PRODUCTS_ENABLED')) {
-            $sql .= "LEFT JOIN ".$this->tablePrefix."barcodes ProductBarcodes ON ProductBarcodes.product_id = Products.id_product ";
-            $sql .= "LEFT JOIN ".$this->tablePrefix."product_attribute ProductAttributes ON ProductAttributes.id_product = Products.id_product ";
-            $sql .= "LEFT JOIN ".$this->tablePrefix."barcodes ProductAttributeBarcodes ON ProductAttributeBarcodes.product_attribute_id = ProductAttributes.id_product_attribute ";
-        }
-
-        return $sql;
-    }
-
-    protected function getOrdersForProductListQuery(): array
-    {
-        return [
-            'Products.name' => 'ASC',
-            'Images.id_image' => 'DESC',
-        ];
-    }
-
-
-    protected function hideIfPurchasePriceNotSet($products)
-    {
-        if (!Configure::read('appDb.FCS_PURCHASE_PRICE_ENABLED')) {
-            return $products;
-        }
-
-        $this->Product = FactoryLocator::get('Table')->get('Products');
-        $i = 0;
-
-        foreach($products as $product) {
-
-            $attributes = $this->Product->ProductAttributes->find('all', [
-                'conditions' => [
-                    'ProductAttributes.id_product' => $product['id_product'],
-                ],
-            ]);
-
-            $attributesCount = $attributes->count();
-
-            if ($attributesCount > 0) {
-
-                foreach($attributes as $attribute) {
-
-                    $attributesWithAssociations = $this->Product->ProductAttributes->find('all', [
-                        'conditions' => [
-                            'ProductAttributes.id_product_attribute' => $attribute->id_product_attribute,
-                        ],
-                        'contain' => [
-                            'PurchasePriceProductAttributes',
-                            'UnitProductAttributes',
-                        ],
-                    ])->first();
-
-                    if (!$this->Product->ProductAttributes->PurchasePriceProductAttributes->isPurchasePriceSet($attributesWithAssociations)) {
-                        $attributesCount--;
-                    }
-
-                }
-
-            }
-
-            if ($attributesCount == 0) {
-                $unitProductEntity = $this->Product->UnitProducts->newEntity([
-                        'price_per_unit_enabled' => $product['price_per_unit_enabled'],
-                        'purchase_price_incl_per_unit' => $product['purchase_price_incl_per_unit'],
-                    ],
-                    [
-                        'validate' => false
-                    ],
-                );
-                $purchasePriceEntity = $this->Product->PurchasePriceProducts->find('all', [
-                    'conditions' => [
-                        'PurchasePriceProducts.product_id' => $product['id_product'],
-                    ],
-                ])->first();
-                $productEntity = [
-                    'unit_product' => $unitProductEntity,
-                    'purchase_price_product' => $purchasePriceEntity,
-                ];
-                $productEntity = json_decode(json_encode($productEntity), false); // convert array recursively into object
-                if (!$this->Product->PurchasePriceProducts->isPurchasePriceSet($productEntity)) {
-                    unset($products[$i]);
-                }
-            }
-
-            $i++;
-        }
-
-        $products = $this->reindexArray($products);
-        return $products;
-    }
-
-    protected function hideProductsWithActivatedDeliveryRhythmOrDeliveryBreak($appAuth, $products)
-    {
-
-        if (Configure::read('appDb.FCS_CUSTOMER_CAN_SELECT_PICKUP_DAY') || $appAuth->isOrderForDifferentCustomerMode() || $appAuth->isSelfServiceModeByUrl()) {
-            return $products;
-        }
-
-        $this->Product = FactoryLocator::get('Table')->get('Products');
-        $i = -1;
-        foreach($products as $product) {
-            $i++;
-            $deliveryDate = $this->Product->calculatePickupDayRespectingDeliveryRhythm($product);
-
-            // deactivates the product if it can not be ordered this week
-            if ($deliveryDate == 'delivery-rhythm-triggered-delivery-break') {
-                $products[$i]->delivery_break_enabled = true;
-            }
-
-            // deactivates the product if manufacturer based delivery break is enabled
-            if ($this->Product->deliveryBreakEnabled($product->manufacturer->no_delivery_days, $deliveryDate)) {
-                $products[$i]->delivery_break_enabled = true;
-            }
-
-            // deactivates the product if global delivery break is enabled
-            if ($this->Product->deliveryBreakEnabled(Configure::read('appDb.FCS_NO_DELIVERY_DAYS_GLOBAL'), $deliveryDate)) {
-                $products[$i]->delivery_break_enabled = true;
-            }
-
-            // hides products when order_possible_until is reached (do not apply if product is stock product)
-            if (!($product->is_stock_product && $product->stock_management_enabled) &&
-                $product->delivery_rhythm_type == 'individual' &&
-                !is_null($product->delivery_rhythm_order_possible_until) &&
-                $product->delivery_rhythm_order_possible_until->i18nFormat(Configure::read('app.timeHelper')->getI18Format('Database')) < Configure::read('app.timeHelper')->getCurrentDateForDatabase()) {
-                    unset($products[$i]);
-            }
-
-        }
-
-        $products = $this->reindexArray($products);
-        return $products;
-
-    }
-
-    protected function reindexArray($array)
-    {
-        $reindexedArray = [];
-        foreach($array as $a) {
-            $reindexedArray[] = $a;
-        }
-        return $reindexedArray;
     }
 
 }
